Renames server/lib as server/src per https://github.com/DoSomething/gateway-js/pull/21/files#r240391939
