+++ server/index.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const GatewayClient = require('./lib/client');
+const GatewayClient = require('./src/client');
 
 module.exports = {
   GatewayClient,
