+++ server/src/endpoints/rogue/signups.js
